% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history-package.R
\docType{package}
\name{blame-package}
\alias{blame}
\alias{blame-package}
\title{blame: Semantic Version Control for R}
\description{
Explore an object's git history.
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
