% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{git_history}
\alias{git_history}
\title{Fetch Branch at the Object Level}
\usage{
git_history(repo = ".", branch = "main")
}
\arguments{
\item{repo}{Path or url to a git repo.}

\item{branch}{Name of the branch of interest.}
}
\description{
This builds a data frame containing the history of all objects of a package
on a given branch.
}
